Deprecate wstat/stat in favour of WinAPI GetFileAttributesEx - unifies code and behaviour between Windows versions
Fixes #4491, fixes #3969, fixes #2535, fixes #215, close #4541

+++ PowerEditor/src/ScitillaComponent/Buffer.h
@@ -310,7 +310,7 @@ public:
 		return _pManager->docLength(_id);
 	}
 
-	int getFileLength() const; // return file length. -1 if file is not existing.
+	int64_t getFileLength() const; // return file length. -1 if file is not existing.
 
 	enum fileTimeType { ft_created, ft_modified, ft_accessed };
 	generic_string getFileTime(fileTimeType ftt) const;
